RING-29369 All returned HTTP streams must be consumed
Erasure coding and replication are not using
all fetched streams to answer the client. If
any of those unused stream is not consumed, it
eventually trigger a timeout and has unfortunate
consequences in upper CloudServer layer.
If CloudServer has not finished to answer (large file)
then the GET will fail (short read/corruption/connection closed, etc).