RING-29695 PUTs should not fail when hyperdrive is under heavy load
When congestion occurs in several hyperdrive modules, it will reject
new PUTs by default to avoid making the problem worse.
Hdclient cannot retry, and hence we go for the following solution:
always force the PUT (adding ?force=true to the url) and instead
lessen the weight of the hyperdrive if under load (we do poll the
statuses regularly).